backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
sqlparse==0.5.5
tzdata==2025.3
# Production dependencies
gunicorn==21.2.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.1
